+++ celery_salt/core/__init__.py
@@ -1,7 +1,7 @@
 """Core CelerySalt functionality."""
 
 from celery_salt.core.decorators import event, subscribe
-from celery_salt.core.events import SaltEvent
+from celery_salt.core.events import SaltEvent, SaltResponse
 from celery_salt.core.exceptions import (
     CelerySaltError,
     PublishError,
@@ -18,6 +18,7 @@ __all__ = [
     "event",
     "subscribe",
     "SaltEvent",
+    "SaltResponse",
     "CelerySaltError",
     "SchemaConflictError",
     "SchemaRegistryUnavailableError",